theme: move pageTheme to root
