+++ packages/theme/src/pageTheme.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { PageTheme } from '../types';
+import { PageTheme } from './types';
 
 /**
  * The default predefined burst shapes.
